final update

+++ semseg_dim_data_gen.py
@@ -190,7 +190,7 @@ class DIMDataset(Dataset):
         y[0, :, :] = alpha / 255.
         # mask = np.equal(trimap, 128).astype(np.float32)
         label_path = self.labels[i]
-        mask = cv.imread(label_path, cv.IMREAD_UNCHANGED)
+        mask = cv.imread(label_path, cv.IMREAD_UNCHANGED).astype(np.float32)
         # mask = safe_crop(mask, x, y, crop_size)
         # mask = np.equal(trimap, 1).astype(np.float32)
         y[1, :, :] = mask
